This script will first determines if a server is on a cluster. If so, it'll check the state of various cluster related objects, including: nodes resource groups, resources, networks and network interfaces.
This script will first determines if a server is on a cluster. If so, it'll check the state of various cluster related objects, including: nodes resource groups, resources, networks and network interfaces. 

NOTE: This script was part of the SVAC R1/R2 solution that used the HPSA WINDOWS SVAC Audit Policy.