This script deletes unnecessary files on the C: drive such as tmp files, dmp files, AV old definitions, old Windows update files, and the Recycle Bin.
Deletes unnecessary files on the C: drive such as tmp files, dmp files, AV old definitions, old Windows update files, and the Recycle Bin.

This batch file releases space on the C: drive on Windows servers according to the standard disk cleanup process of the Clorox customer.

This script needs to be saved in a folder named C:\script. A file named wintel-users.txt with all Wintel user names needs to be located within this folder.