This script is used to create a server status report after the server has an unexpected reboot or to verify the server when you want.
This script is used to create a server status report after the server has an unexpected reboot or to verify the server when you want.

The Output is a status of the server including Memory, Virtual memory, System Events (last three boots, HW errors, Patch Installing), Logical Disk, Shared, Printer, Services, Process, Network and Active Directory (if server is Domain Controller).

The script output shows only errors and the general status. The script has a parameter (/verbose) to expand the output to show all information, not only that have errors.